feat(pragma): add optInOnly option
