+++ packages/babel-plugin-flow-runtime/src/ConversionContext.js
@@ -30,6 +30,7 @@ export default class ConversionContext {
   shouldAssert: boolean = true;
   shouldWarn: boolean = false;
   shouldAnnotate: boolean = true;
+  optInOnly: boolean = false;
   isAnnotating: boolean = false;
   suppressCommentPatterns: RegExp[] = [/\$FlowFixMe/];
   suppressTypeNames: string[] = ['$FlowFixMe'];
